Refactor: separate version upgrade action saga
